move: lotto 관련 클래스의 패키지 경로 이동
