+++ src/main/java/lotto/domain/WinningLotto.java
@@ -3,6 +3,8 @@ package lotto.domain;
 import static lotto.constant.ExceptionMessage.BONUS_NUMBER_DUPLICATED_WITH_WINNING_NUMBER;
 
 import lotto.constant.Rank;
+import lotto.domain.lotto.Lotto;
+import lotto.domain.lotto.Number;
 
 public class WinningLotto {
