+++ src/main/java/lotto/dto/response/LottosResponse.java
@@ -2,7 +2,7 @@ package lotto.dto.response;
 
 import java.util.List;
 
-import lotto.domain.Lotto;
+import lotto.domain.lotto.Lotto;
 
 public record LottosResponse(
     List<InnerLottoResponse> lottos
