+++ src/test/java/lotto/domain/LottoTest.java
@@ -11,6 +11,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 import lotto.constant.Rank;
+import lotto.domain.lotto.Lotto;
 import lotto.random.LottoRandom;
 
 class LottoTest {
